pyproject.toml
[tool.poetry]
name = "ai-shopping-concierge"
version = "0.1.0"
description = ""
authors = ["Yushi Homma <hommayushi3@gmail.com>"]
readme = "README.md"
packages = [
    { include = "realtime", from = "." },
]

[tool.poetry.dependencies]
python = "^3.11"
chainlit = "^2.0.dev2"
openai = "*"
websockets = "^13.1"
pillow = "^10.4.0"
litellm = "^1.48.18"
pinecone-client = "^5.0.1"
cohere = "^5.11.1"
pandas = "^2.2.3"
google-generativeai = "^0.8.3"
sse-starlette = "^2.1.3"
pydantic = "^2.9.2"
redis = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
